+++ tencentcloud/resource_tc_instance.go
@@ -157,7 +157,7 @@ func resourceTencentCloudInstance() *schema.Resource {
 				ForceNew:     true,
 				Default:      CVM_CHARGE_TYPE_POSTPAID,
 				ValidateFunc: validateAllowedStringValue(CVM_CHARGE_TYPE),
-				Description:  "The charge type of instance. Valid values are `PREPAID`, `POSTPAID_BY_HOUR` and `SPOTPAID`, The default is `POSTPAID_BY_HOUR`.",
+				Description:  "The charge type of instance. Valid values are `PREPAID`, `POSTPAID_BY_HOUR` and `SPOTPAID`, The default is `POSTPAID_BY_HOUR`. Note: TencentCloud International only supports `POSTPAID_BY_HOUR`, `PREPAID` instance may not allow to delete before expired.",
 			},
 			"instance_charge_type_prepaid_period": {
 				Type:         schema.TypeInt,